[List v2]: Handle Enter in empty list items
